support folder delete, prevent ghs.yml see

+++ assets/js/index.js
@@ -251,8 +251,10 @@ var vm = new Vue({
       })
     },
     deletePathConfirm: function (f, e) {
-      // confirm
       e.preventDefault();
+      if (!window.confirm("Delete " + f.name + " ?")) {
+        return;
+      }
       $.ajax({
         url: pathJoin([location.pathname, f.name]),
         method: 'DELETE',
